[Nearby Share] Fix transfer size truncation
Fixes an issue where the transfer size in the
CUJ metric is truncated to ~4gb.

(cherry picked from commit ae195ef4c40dea2cfc6aabc954a7bf46c07e531a)

Test: Verified locally
Bug: b/313520497
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5126552
Auto-Submit: Jack Shira <jackshira@google.com>
Commit-Queue: Jack Shira <jackshira@google.com>
Commit-Queue: Daniel Classon <dclasson@google.com>
Cr-Original-Commit-Position: refs/heads/main@{#1238173}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5126301
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/branch-heads/6167@{#409}
Cr-Branched-From: 222e786949e76e342d325ea0d008b4b6273f3a89-refs/heads/main@{#1233107}

+++ chrome/browser/nearby_sharing/metrics/nearby_share_metric_logger.h
@@ -52,7 +52,7 @@ class NearbyShareMetricLogger : public NearbySharingService::Observer {
       share_target_upgrade_time_;
   base::flat_map<base::UnguessableToken, nearby::connections::mojom::Medium>
       share_target_medium_;
-  base::flat_map<base::UnguessableToken, uint> transfer_size_;
+  base::flat_map<base::UnguessableToken, int64_t> transfer_size_;
   base::flat_map<base::UnguessableToken, float> transfer_progress_;
 };
